[FTR] Async Task GCE Communication
Feature Description

The async task is now being launched from the main activity when the
user presses the `Tell Joke` button.  This will launch the task to query
for a string from the GCE service.

At this point in time however, returning error string of `Cannot connect
to localhost`

Related Issues